Move all components into their own folders
