+++ frontend/src/components/Account/Login.jsx
@@ -1,20 +1,19 @@
-import "./Register.css"
+import "./Login.css"
 import { useState } from "react"
 import { toast } from "react-toastify"
-import { IoPersonAdd } from "react-icons/io5"
+import { FaSignInAlt } from "react-icons/fa"
 import axios from "axios"
-import Spinner from "../components/Spinner"
+import Spinner from "../Spinner/Spinner"
 
-const Register = (props) => {
+const Login = (props) => {
     const Form = () => {
         const [formData, setFormData] = useState({
             name: "",
             password: "",
-            password2: "",
         })
         const [loading, setLoading] = useState(false)
 
-        const { name, password, password2 } = formData
+        const { name, password } = formData
 
         const onChange = (e) => {
             setFormData((prevState) => ({
@@ -25,20 +24,17 @@ const Register = (props) => {
 
         const onSubmit = (e) => {
             e.preventDefault()
-            if (name === "" || password === "" || password2 === "") {
+
+            if (name === "" || password === "") {
                 toast.error("Please fill in all fields")
-            } else if (name.length > 16)
-                toast.error("Username exceeds 16 character limit")
-            else if (password !== password2)
-                toast.error("Passwords do not match")
-            else {
+            } else {
                 setLoading(true)
                 const userData = {
                     name,
                     password,
                 }
                 axios
-                    .post("/api/users", userData)
+                    .post("/api/users/login", userData)
                     .then((response) => {
                         if (response.data) {
                             localStorage.setItem(
@@ -46,7 +42,6 @@ const Register = (props) => {
                                 JSON.stringify(response.data)
                             )
                             props.setLoggedIn(true)
-                            props.setToggleRegister(false)
                         }
                         setLoading(false)
                     })
@@ -82,35 +77,18 @@ const Register = (props) => {
                     />
                 </div>
                 <div className="form-item">
-                    <input
-                        required
-                        type="password"
-                        id="password2"
-                        name="password2"
-                        value={password2}
-                        placeholder="Confirm Password"
-                        onChange={onChange}
-                    />
-                </div>
-                <div className="form-item">
                     {loading ? (
                         <div className="spinner-container">
                             <Spinner size="small" />
                         </div>
                     ) : (
-                        <button className="green" type="submit">
-                            Sign Up
-                        </button>
+                        <button type="submit">Sign In</button>
                     )}
                 </div>
                 <hr className="divider" />
                 <div className="form-item">
-                    <button
-                        onClick={() => {
-                            props.setToggleRegister()
-                        }}
-                    >
-                        Login
+                    <button className="green" onClick={props.setToggleRegister}>
+                        Register
                     </button>
                 </div>
             </form>
@@ -118,13 +96,13 @@ const Register = (props) => {
     }
 
     return (
-        <section className="register-container">
+        <section className="login-container">
             <div className="form-title">
-                <IoPersonAdd /> Register
+                <FaSignInAlt /> Login
             </div>
             <Form />
         </section>
     )
 }
 
-export default Register
+export default Login
