+++ frontend/src/components/Compare/Compare.css
