+++ frontend/src/components/Versus/Leaderboard.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react"
 import axios from "axios"
 import { toast } from "react-toastify"
 
-import Spinner from "./Spinner"
+import Spinner from "../Spinner/Spinner"
 
 import { AiOutlineClose } from "react-icons/ai"
 import { FaCrown } from "react-icons/fa"
