+++ frontend/src/components/Versus/Versus.jsx
@@ -7,9 +7,9 @@ import {
 } from "react-circular-progressbar"
 import "react-circular-progressbar/dist/styles.css"
 
-import { AppContext } from "../AppContext"
-import roster from "../players.json"
-import { teamColors, defaultColor } from "../config"
+import { AppContext } from "../../AppContext"
+import roster from "../../players.json"
+import { teamColors, defaultColor } from "../../config"
 import Stats from "./Stats"
 import Leaderboard from "./Leaderboard"
